finish css  + refactor
Header check all css
Small refactor
No validation on edit profile form
No Loader
