+++ src/components/header/Modal/Modal.module.css
@@ -12,18 +12,23 @@
   width: 28px;
   stroke: rgba(34, 37, 42, 1);
   background-color: transparent;
-  &:hover {
+  &:hover,
+  &:focus {
     scale: 1.2;
     stroke: rgba(139, 170, 54, 1);
   }
 }
+/* .relative {
+  position: relative;
+} */
 
 .modalOverlay {
   position: absolute;
   top: 0;
   left: 0;
-  width: 100%;
-  height: 100%;
+  min-width: 375px;
+  width: 100vw;
+  height: 100vh;
   background-color: rgba(235, 243, 212, 1);
   background-image: url('../../../images/bgc/menu_mob_1x.png');
   background-repeat: no-repeat;
