+++ src/components/header/UserProfile/UserLogo.jsx
@@ -3,7 +3,7 @@ import UserLogoModal from './UserLogoModal/UserLogoModal';
 import { useDispatch, useSelector } from 'react-redux';
 import { selectUserAvatar, selectUserName } from 'redux/auth/selectors';
 import css from './UserProfile.module.css';
-import defaultUserAvatar from '../../../pictures/userDefault.png';
+import defaultUserAvatar from '../../../images/header/userDefault.png';
 import UserLogoutModal from './UserLogoutModal';
 import UserInfoModal from './UserInfoModal/UserInfoModal';
 import { refreshUser } from 'redux/auth/operations';
