+++ src/components/header/UserProfile/UserLogoModal/UserLogo.module.css
@@ -1,7 +1,7 @@
 .UserLogoModal {
   position: absolute;
 
-  top: 11%;
+  top: 40px;
   left: 60%;
   width: 177px;
   height: 134px;
@@ -94,7 +94,7 @@
 }
 @media screen and (min-width: 768px) {
   .UserLogoModal {
-    top: 8%;
+    top: 80px;
     left: 50%;
     gap: 32px;
   }
