+++ src/pages/pages.module.css
@@ -1,32 +1,30 @@
-  @media screen and (min-width: 375px) {
-        .section {
-            padding-bottom: 50px;
-            
-            
-        }
-        .container{
-            width: 375px;
-            padding: 18px 16px 0 16px;
-            margin: 0 auto;
-            
-        }
-    }
-    @media screen and (min-width: 768px) {
-        .section {
-            padding-bottom: 40px;
-        }
-        .container{
-          width: 768px;  
-          padding: 18px 32px 0 32px;
-        }
-    }
-  
-    @media screen and (min-width: 1440px) {
-      .section {
-        padding-bottom: 50px;
-       }
-       .container{  
-           width: 1440px;
-           padding: 18px 100px 0 100px;
-        }
-    }
+@media screen and (min-width: 375px) {
+  .section {
+    padding-bottom: 50px;
+  }
+  .container {
+    min-width: 375px;
+    width: 100%;
+    padding: 18px 16px 0 16px;
+    margin: 0 auto;
+  }
+}
+@media screen and (min-width: 768px) {
+  .section {
+    padding-bottom: 40px;
+  }
+  .container {
+    width: 768px;
+    padding: 18px 32px 0 32px;
+  }
+}
+
+@media screen and (min-width: 1440px) {
+  .section {
+    padding-bottom: 50px;
+  }
+  .container {
+    width: 1440px;
+    padding: 18px 100px 0 100px;
+  }
+}
